requirements.txt
Django==1.6.2
Django-Select2==4.2.2
Minidetector==0.1
PyYAML==3.10
dj-database-url==0.3.0
dj-static==0.0.5
django-angular==0.6.3
django-appconf==0.6
django-bootstrap-form==3.1
django-compressor==1.3
django-cors-headers==0.12
django-heroku-memcacheify==0.5
django-pylibmc-sasl==0.2.4
django-timezone-field==1.0
gunicorn==18.0
psycopg2==2.5.2
pygeoip==0.3.1
pystache==0.5.3
pytz==2014.2
six==1.6.1
static==1.0.2
ua-parser==0.3.4
